+++ packages/permission/src/domain/index.ts
@@ -1,1 +1,2 @@
 export * from "./permission-engine";
+export * from "./permission-ast-transform";